View portfolio security: New test created-> New buy action leads to portfolio change

+++ WealthWise/src/main/java/com/bourse/wealthwise/domain/entity/action/Buy.java
@@ -16,7 +16,7 @@ import java.util.UUID;
 public class Buy extends BaseAction {
     private final Security security;
     private final BigInteger volume;
-    private final Integer price;
+    private final Double price;
     private final BigInteger totalValue;
 
     @Override
